+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 * Add the module `pyctr.type.configsave` with the class `ConfigSaveReader`
+* Implement `to_bytes` and `remove_block` in `ConfigSaveReader`
 
 ## v0.6.0 - January 26, 2022
 ### Highlights